+++ doc/api/search/all_0.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['addfactory',['AddFactory',['../class_pip_services_1_1_container_1_1_container.html#a35cd54f1252b5a51eaa68e0a35def651',1,'PipServices::Container::Container']]]
+  ['addfactory',['AddFactory',['../class_pip_services_1_1_container_1_1_container.html#a35cd54f1252b5a51eaa68e0a35def651',1,'PipServices3::Container::Container']]]
 ];