+++ doc/api/search/all_1.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['basereferences',['BaseReferences',['../class_pip_services_1_1_container_1_1_refer_1_1_references_decorator.html#a830c2a71494c103c767bad7a16fd6703',1,'PipServices::Container::Refer::ReferencesDecorator']]],
-  ['buildreferencesdecorator',['BuildReferencesDecorator',['../class_pip_services_1_1_container_1_1_refer_1_1_build_references_decorator.html',1,'PipServices.Container.Refer.BuildReferencesDecorator'],['../class_pip_services_1_1_container_1_1_refer_1_1_build_references_decorator.html#ab770b8e5026a325be5f37177980c438d',1,'PipServices.Container.Refer.BuildReferencesDecorator.BuildReferencesDecorator()']]]
+  ['basereferences',['BaseReferences',['../class_pip_services_1_1_container_1_1_refer_1_1_references_decorator.html#a830c2a71494c103c767bad7a16fd6703',1,'PipServices3::Container::Refer::ReferencesDecorator']]],
+  ['buildreferencesdecorator',['BuildReferencesDecorator',['../class_pip_services_1_1_container_1_1_refer_1_1_build_references_decorator.html',1,'PipServices3.Container.Refer.BuildReferencesDecorator'],['../class_pip_services_1_1_container_1_1_refer_1_1_build_references_decorator.html#ab770b8e5026a325be5f37177980c438d',1,'PipServices3.Container.Refer.BuildReferencesDecorator.BuildReferencesDecorator()']]]
 ];