+++ doc/api/search/all_a.js
@@ -1,12 +1,12 @@
 var searchData=
 [
-  ['build',['Build',['../namespace_pip_services_1_1_container_1_1_build.html',1,'PipServices::Container']]],
-  ['config',['Config',['../namespace_pip_services_1_1_container_1_1_config.html',1,'PipServices::Container']]],
-  ['container',['Container',['../namespace_pip_services_1_1_container.html',1,'PipServices']]],
-  ['parentreferences',['ParentReferences',['../class_pip_services_1_1_container_1_1_refer_1_1_references_decorator.html#a4db9cc6133e9ad682ca260bc4f27f0d1',1,'PipServices::Container::Refer::ReferencesDecorator']]],
-  ['pipservices',['PipServices',['../namespace_pip_services.html',1,'']]],
-  ['processcontainer',['ProcessContainer',['../class_pip_services_1_1_container_1_1_process_container.html',1,'PipServices.Container.ProcessContainer'],['../class_pip_services_1_1_container_1_1_process_container.html#ac66d80eabc20b055a0e076696a5b7e45',1,'PipServices.Container.ProcessContainer.ProcessContainer()']]],
-  ['put',['Put',['../class_pip_services_1_1_container_1_1_refer_1_1_link_references_decorator.html#a3d3f180d7356853acf996582259107f2',1,'PipServices.Container.Refer.LinkReferencesDecorator.Put()'],['../class_pip_services_1_1_container_1_1_refer_1_1_references_decorator.html#a37fdba90d393892047ca79b11481a58a',1,'PipServices.Container.Refer.ReferencesDecorator.Put()'],['../class_pip_services_1_1_container_1_1_refer_1_1_run_references_decorator.html#a3e024d99afa1ac42f3f346cb3730af57',1,'PipServices.Container.Refer.RunReferencesDecorator.Put()']]],
-  ['putfromconfig',['PutFromConfig',['../class_pip_services_1_1_container_1_1_refer_1_1_container_references.html#aad7657e69928ce1e9f26b8644a52f6a4',1,'PipServices::Container::Refer::ContainerReferences']]],
-  ['refer',['Refer',['../namespace_pip_services_1_1_container_1_1_refer.html',1,'PipServices::Container']]]
+  ['build',['Build',['../namespace_pip_services_1_1_container_1_1_build.html',1,'PipServices3::Container']]],
+  ['config',['Config',['../namespace_pip_services_1_1_container_1_1_config.html',1,'PipServices3::Container']]],
+  ['container',['Container',['../namespace_pip_services_1_1_container.html',1,'PipServices3']]],
+  ['parentreferences',['ParentReferences',['../class_pip_services_1_1_container_1_1_refer_1_1_references_decorator.html#a4db9cc6133e9ad682ca260bc4f27f0d1',1,'PipServices3::Container::Refer::ReferencesDecorator']]],
+  ['pipservices',['PipServices3',['../namespace_pip_services.html',1,'']]],
+  ['processcontainer',['ProcessContainer',['../class_pip_services_1_1_container_1_1_process_container.html',1,'PipServices3.Container.ProcessContainer'],['../class_pip_services_1_1_container_1_1_process_container.html#ac66d80eabc20b055a0e076696a5b7e45',1,'PipServices3.Container.ProcessContainer.ProcessContainer()']]],
+  ['put',['Put',['../class_pip_services_1_1_container_1_1_refer_1_1_link_references_decorator.html#a3d3f180d7356853acf996582259107f2',1,'PipServices3.Container.Refer.LinkReferencesDecorator.Put()'],['../class_pip_services_1_1_container_1_1_refer_1_1_references_decorator.html#a37fdba90d393892047ca79b11481a58a',1,'PipServices3.Container.Refer.ReferencesDecorator.Put()'],['../class_pip_services_1_1_container_1_1_refer_1_1_run_references_decorator.html#a3e024d99afa1ac42f3f346cb3730af57',1,'PipServices3.Container.Refer.RunReferencesDecorator.Put()']]],
+  ['putfromconfig',['PutFromConfig',['../class_pip_services_1_1_container_1_1_refer_1_1_container_references.html#aad7657e69928ce1e9f26b8644a52f6a4',1,'PipServices3::Container::Refer::ContainerReferences']]],
+  ['refer',['Refer',['../namespace_pip_services_1_1_container_1_1_refer.html',1,'PipServices3::Container']]]
 ];