+++ doc/api/search/namespaces_0.js
@@ -1,8 +1,8 @@
 var searchData=
 [
-  ['build',['Build',['../namespace_pip_services_1_1_container_1_1_build.html',1,'PipServices::Container']]],
-  ['config',['Config',['../namespace_pip_services_1_1_container_1_1_config.html',1,'PipServices::Container']]],
-  ['container',['Container',['../namespace_pip_services_1_1_container.html',1,'PipServices']]],
-  ['pipservices',['PipServices',['../namespace_pip_services.html',1,'']]],
-  ['refer',['Refer',['../namespace_pip_services_1_1_container_1_1_refer.html',1,'PipServices::Container']]]
+  ['build',['Build',['../namespace_pip_services_1_1_container_1_1_build.html',1,'PipServices3::Container']]],
+  ['config',['Config',['../namespace_pip_services_1_1_container_1_1_config.html',1,'PipServices3::Container']]],
+  ['container',['Container',['../namespace_pip_services_1_1_container.html',1,'PipServices3']]],
+  ['pipservices',['PipServices3',['../namespace_pip_services.html',1,'']]],
+  ['refer',['Refer',['../namespace_pip_services_1_1_container_1_1_refer.html',1,'PipServices3::Container']]]
 ];