+++ src/Refer/LinkReferencesDecorator.cs
@@ -1,14 +1,14 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
-using PipServices.Commons.Refer;
-using PipServices.Commons.Run;
+using PipServices3.Commons.Refer;
+using PipServices3.Commons.Run;
 
-namespace PipServices.Container.Refer
+namespace PipServices3.Container.Refer
 {
     /// <summary>
     /// References decorator that automatically sets references to newly added components
-    /// that implement <a href="https://rawgit.com/pip-services-dotnet/pip-services-commons-dotnet/master/doc/api/interface_pip_services_1_1_commons_1_1_refer_1_1_i_referenceable.html">IReferenceable</a> interface and unsets references from removed components
-    /// that implement <a href="https://rawgit.com/pip-services-dotnet/pip-services-commons-dotnet/master/doc/api/interface_pip_services_1_1_commons_1_1_refer_1_1_i_unreferenceable.html">IUnreferenceable</a> interface.
+    /// that implement <a href="https://rawgit.com/pip-services3-dotnet/pip-services3-commons-dotnet/master/doc/api/interface_pip_services_1_1_commons_1_1_refer_1_1_i_referenceable.html">IReferenceable</a> interface and unsets references from removed components
+    /// that implement <a href="https://rawgit.com/pip-services3-dotnet/pip-services3-commons-dotnet/master/doc/api/interface_pip_services_1_1_commons_1_1_refer_1_1_i_unreferenceable.html">IUnreferenceable</a> interface.
     /// </summary>
     public class LinkReferencesDecorator : ReferencesDecorator, IOpenable
     {