+++ src/Refer/ReferencesDecorator.cs
@@ -1,11 +1,11 @@
-﻿using PipServices.Commons.Refer;
-using PipServices.Commons.Reflect;
+﻿using PipServices3.Commons.Refer;
+using PipServices3.Commons.Reflect;
 using System.Collections.Generic;
 
-namespace PipServices.Container.Refer
+namespace PipServices3.Container.Refer
 {
     /// <summary>
-    /// Chainable decorator for <a href="https://rawgit.com/pip-services-dotnet/pip-services-commons-dotnet/master/doc/api/interface_pip_services_1_1_commons_1_1_refer_1_1_i_references.html">IReferences</a> that allows to inject additional capabilities
+    /// Chainable decorator for <a href="https://rawgit.com/pip-services3-dotnet/pip-services3-commons-dotnet/master/doc/api/interface_pip_services_1_1_commons_1_1_refer_1_1_i_references.html">IReferences</a> that allows to inject additional capabilities
     /// such as automatic component creation, automatic registration and opening.
     /// </summary>
     /// See <see cref="IReferences"/>